ajout suppression de son compte, dashboard avec utilisateurs et autorisations, apicontroller pour un rate-limiter et avancement du front

+++ src/Controller/SessionFormationController.php
@@ -28,7 +28,7 @@ class SessionFormationController extends AbstractController
     // ---------------------------------------------------------- session --------------------------------------------
 
     //liste des sessions
-    #[Route('/session', name: 'app_session')]
+    #[Route('/', name: 'app_session')] //chemin 'path', '/' pour que l'application s'ouvre directement avec ça
     public function index(SessionRepository $sessionRepository): Response
     //faire passer directement en argument le repository pour appeler des méthodes plus rapidement
     {
@@ -40,38 +40,43 @@ class SessionFormationController extends AbstractController
     }
 
     //crée une nouvelle session
-    #[IsGranted(new Expression('is_granted("ROLE_USER")'))] //autorisation pour effectuer certaines fonctionnalités
+    // #[IsGranted(new Expression('is_granted("ROLE_USER")'))] //autorisation pour effectuer certaines fonctionnalités
     #[Route('/session/newSession', name: 'new_session')]
     public function newSession(SessionRepository $sessionRepository, EntityManagerInterface $entityManager, Request $request)
     {
-        
-        $session = new Session();
+        if ($this->isGranted('ROLE_USER')){
 
-        //crée le form
-        $form = $this->createForm(SessionType::class, $session);
-
-        //prend en charge
-        $form->handleRequest($request);
-
-        if($form->isSubmitted() && $form->isValid()){
-            //récupère les données du formulaire 
-            $session = $form->getData();
-            
-            //verifie la cohérence des dates, fonction dans l'entité qui renvoie un boolean
-            if($session->getValidDate()){
-                $entityManager->persist($session); //prepare
-                $entityManager->flush(); //execute
+            $session = new Session();
     
-                //notif et redirige vers la liste des sessions
-                $this->addFlash('success', 'Session créée');
-                return $this->redirectToRoute("app_session");
-            } else {
-                $this->addFlash('error', 'La date de début est soit déjà passée, soit après la date de fin');
-                return $this->redirectToRoute("app_session") ;
+            //crée le form
+            $form = $this->createForm(SessionType::class, $session);
+    
+            //prend en charge
+            $form->handleRequest($request);
+    
+            if($form->isSubmitted() && $form->isValid()){
+                //récupère les données du formulaire 
+                $session = $form->getData();
+                
+                //verifie la cohérence des dates, fonction dans l'entité qui renvoie un boolean
+                if($session->getValidDate()){
+                    $entityManager->persist($session); //prepare
+                    $entityManager->flush(); //execute
+        
+                    //notif et redirige vers la liste des sessions
+                    $this->addFlash('success', 'Session créée');
+                    return $this->redirectToRoute("app_session");
+                } else {
+                    $this->addFlash('error', 'La date de début est soit déjà passée, soit après la date de fin');
+                    return $this->redirectToRoute("app_session") ;
+                }
             }
-
+            
+        } else {
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
         }
-
+        
         //renvoie la vue
         return $this->render('session/new.html.twig', [
             'form' => $form,
@@ -85,69 +90,84 @@ class SessionFormationController extends AbstractController
     //modifie la session
     #[Route('/session/{id}/edit', name: 'edit_session')] 
     public function edit(Session $session, SessionRepository $sessionRepository, EntityManagerInterface $entityManager, Request $request){
-        //crée le form
-        $form = $this->createForm(EditSessionType::class, $session);
-    
-        //prend en charge
-        $form->handleRequest($request);
-    
-        if($form->isSubmitted() && $form->isValid()){
-            //récupère les données du formulaire 
-            $session = $form->getData();
 
-            $entityManager->persist($session); //prepare
-            $entityManager->flush(); //execute
-
-            //notif et redirige vers le detail de la session
-            $this->addFlash('success', 'Session bien modifiée');
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);         
+        if ($this->isGranted('ROLE_USER')){
             
+            //crée le form
+            $form = $this->createForm(EditSessionType::class, $session);
+        
+            //prend en charge
+            $form->handleRequest($request);
+        
+            if($form->isSubmitted() && $form->isValid()){
+                //récupère les données du formulaire 
+                $session = $form->getData();
+    
+                $entityManager->persist($session); //prepare
+                $entityManager->flush(); //execute
+    
+                //notif et redirige vers le detail de la session
+                $this->addFlash('success', 'Session bien modifiée');
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);      
+            }
+
+            //renvoie la vue
+            return $this->render('session/edit.html.twig', [
+                'form' => $form,
+                'sessionId' => $session->getId()
+            ]);
+
+        } else {
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
         }
-    
-        //renvoie la vue
-        return $this->render('session/edit.html.twig', [
-            'form' => $form,
-            'sessionId' => $session->getId()
-        ]);
-    
+            
     }
 
     //ajoute un stagiaire à la session
     #[Route('/session/{idSession}/addStagiaireSession/{idStagiaire}', name: 'add_stagiaire_session')] 
     public function addStagiaireSession(SessionRepository $sr, EntityManagerInterface $entityManager, StagiaireRepository $stagiaireR ,Request $request, int $idStagiaire, int $idSession) : Response
     {
-        $stagiaire = $stagiaireR->findOneById($idStagiaire); //objet stagiaire
-        $session = $sr->findOneById($idSession); //objet session
+        //si l'user a l'autorisation
+        if ($this->isGranted('ROLE_USER')){
+            $stagiaire = $stagiaireR->findOneById($idStagiaire); //objet stagiaire
+            $session = $sr->findOneById($idSession); //objet session
         
-        //s'il ne restait plus qu'une place, on ajoute le stagiaire et on ferme l'inscription
-        if($session->getNbDispo() ==1){
-            $session->setOuvert(false); //ferme la session
-            
-            //cette methode dans la classe session attend en argument le stagiaire
-            $add = $session->addInscription($stagiaire);
-
-            $entityManager->persist($add); //prepare
-            $entityManager->flush(); //execute
-
-            //notif redirige vers le detail de la session
-            $this->addFlash('success', 'Stagiaire '. $stagiaire .' ajouté à la session, session fermée' );
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
-
-        } elseif($session->getNbDispo() > 1){
-            //cette methode dans la classe session attend en argument le stagiaire
-            $add = $session->addInscription($stagiaire);
-
-            $entityManager->persist($add); //prepare
-            $entityManager->flush(); //execute
-            
-            //notif redirige vers le detail de la session
-            $this->addFlash('success', 'Stagiaire '. $stagiaire .' ajouté à la session' );
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
-            
+            //s'il ne restait plus qu'une place, on ajoute le stagiaire et on ferme l'inscription
+            if($session->getNbDispo() ==1){
+                $session->setOuvert(false); //ferme la session
+                
+                //cette methode dans la classe session attend en argument le stagiaire
+                $add = $session->addInscription($stagiaire);
+    
+                $entityManager->persist($add); //prepare
+                $entityManager->flush(); //execute
+    
+                //notif redirige vers le detail de la session
+                $this->addFlash('success', 'Stagiaire '. $stagiaire .' ajouté à la session, session fermée' );
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+                
+            } elseif($session->getNbDispo() > 1){
+                //cette methode dans la classe session attend en argument le stagiaire
+                $add = $session->addInscription($stagiaire);
+    
+                $entityManager->persist($add); //prepare
+                $entityManager->flush(); //execute
+                
+                //notif redirige vers le detail de la session
+                $this->addFlash('success', 'Stagiaire '. $stagiaire .' ajouté à la session' );
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+                
+            } else {
+                $this->addFlash('error', 'Session complète');
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+            }
+                
         } else {
-            $this->addFlash('error', 'Session complète');
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
         }
+        
 
         
     }
@@ -156,41 +176,46 @@ class SessionFormationController extends AbstractController
     #[Route('/session/{idSession}/removeStagiaireSession/{idStagiaire}', name: 'remove_stagiaire_session')] 
     public function removeStagiaireSession(SessionRepository $sr, EntityManagerInterface $entityManager, StagiaireRepository $stagiaireR ,Request $request, int $idStagiaire, int $idSession) : Response
     {
-        
-        $stagiaire = $stagiaireR->findOneById($idStagiaire); //objet stagiaire
-        $session = $sr->findOneById($idSession); //objet session
-        
-        //si on enleve un stagiaire alors qu'avant elle était fermée
-        if($session->getNbDispo() ==0){
-            $session->setOuvert(true); //rouvre la session
+        if ($this->isGranted('ROLE_USER')){
 
-            //cette methode dans la classe session attend en argument le stagiaire
-            $delete = $session->removeInscription($stagiaire);
+            $stagiaire = $stagiaireR->findOneById($idStagiaire); //objet stagiaire
+            $session = $sr->findOneById($idSession); //objet session
+            
+            //si on enleve un stagiaire alors qu'avant elle était fermée
+            if($session->getNbDispo() ==0){
+                $session->setOuvert(true); //rouvre la session
     
-            $entityManager->persist($delete); //prepare
-            $entityManager->flush(); //execute
-
-
-            //notif redirige vers le detail de la session
-            $this->addFlash('success', 'Stagiaire ' . $stagiaire . ' supprimé de la session');
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
-
-        } elseif($session->getNbDispo() > 0){
-            //cette methode dans la classe session attend en argument le stagiaire
-            $delete = $session->removeInscription($stagiaire);
+                //cette methode dans la classe session attend en argument le stagiaire
+                $delete = $session->removeInscription($stagiaire);
+        
+                $entityManager->persist($delete); //prepare
+                $entityManager->flush(); //execute
     
-            $entityManager->persist($delete); //prepare
-            $entityManager->flush(); //execute
+    
+                //notif redirige vers le detail de la session
+                $this->addFlash('success', 'Stagiaire ' . $stagiaire . ' supprimé de la session');
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+    
+            } elseif($session->getNbDispo() > 0){
+                //cette methode dans la classe session attend en argument le stagiaire
+                $delete = $session->removeInscription($stagiaire);
+        
+                $entityManager->persist($delete); //prepare
+                $entityManager->flush(); //execute
+    
+                //notif redirige vers le detail de la session
+                $this->addFlash('success', 'Stagiaire ' . $stagiaire . ' supprimé de la session');
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+            } else {
+    
+                return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+            }
 
-            //notif redirige vers le detail de la session
-            $this->addFlash('success', 'Stagiaire ' . $stagiaire . ' supprimé de la session');
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
         } else {
-
-            return $this->redirectToRoute('show_session', ['id'=>$session->getId()]);
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
         }
-        
-        
+            
     }
 
 
@@ -213,13 +238,19 @@ class SessionFormationController extends AbstractController
     #[Route('/session/{id}/deleteSession', name:'delete_session')]
     public function deleteSession(Session $session, EntityManagerInterface $entityManager){
 
+        if($this->isGranted('ROLE_ADMIN')){
+            $entityManager->remove($session);
+            $entityManager->flush();
+    
+            //notif et redirection
+            $this->addFlash('succes', 'Session supprimée');
+            return $this->redirectToRoute('app_session');
+        
+        } else {
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
+        }
 
-        $entityManager->remove($session);
-        $entityManager->flush();
-
-        //notif et redirection
-        $this->addFlash('succes', 'Session supprimée');
-        return $this->redirectToRoute('app_session');
     }
 
 
@@ -259,31 +290,37 @@ class SessionFormationController extends AbstractController
     #[Route('/formation/newFormation', name: 'new_formation')]
     public function newFormation(FormationRepository $formationRepository, EntityManagerInterface $entityManager, Request $request)
     {
-        $formation = new Formation();
+        if($this->isGranted('ROLE_USER')){
 
+            $formation = new Formation();
 
-        //crée le form
-        $form = $this->createForm(FormationType::class, $formation);
+            //crée le form
+            $form = $this->createForm(FormationType::class, $formation);
+    
+            //prend en charge
+            $form->handleRequest($request);
+    
+            if($form->isSubmitted() && $form->isValid()){
+                //récupère les données du formulaire 
+                $formation = $form->getData();
+    
+                $entityManager->persist($formation); //prepare
+                $entityManager->flush(); //execute
+    
+                //notif et redirige vers la liste des formations
+                $this->addFlash('success', 'Formation créée');
+                return $this->redirectToRoute("app_formation");
+            }
+    
+            //renvoie la vue
+            return $this->render('formation/new.html.twig', [
+                'form' => $form
+            ]);
 
-        //prend en charge
-        $form->handleRequest($request);
-
-        if($form->isSubmitted() && $form->isValid()){
-            //récupère les données du formulaire 
-            $formation = $form->getData();
-
-            $entityManager->persist($formation); //prepare
-            $entityManager->flush(); //execute
-
-            //notif et redirige vers la liste des formations
-            $this->addFlash('success', 'Formation créée');
-            return $this->redirectToRoute("app_formation");
+        } else {
+            $this->addFlash('error', 'Vous n\'avez pas l\'autorisation d\'effectuer cette action.');
+            $this->redirectToRoute('app_session');
         }
-
-        //renvoie la vue
-        return $this->render('formation/new.html.twig', [
-            'form' => $form
-        ]);
 
     }
